requirements.txt
fastapi

uvicorn

sqlalchemy

psycopg2-binary

pydantic

alembic

pandas

numpy

scikit-learn

joblib

python-dotenv

streamlit

requests